+++ kermit/k95/ckuus5.c
@@ -6250,6 +6250,9 @@ shotrm() {
       case XYB_AUD | XYB_BEEP:
         ckstrncat(bell,"beep",64);
         break;
+      case XYB_AUD | XYB_MIDI:
+        ckstrncat(bell, "sound device",64);
+        break;
       case XYB_AUD | XYB_SYS:
         ckstrncat(bell,"system sounds",64);
         break;